Move InteractionConstraints to kt source
