+++ src/main/kotlin/com/freya02/botcommands/api/new_components/builder/IConstrainableComponent.kt
@@ -1,7 +1,7 @@
 package com.freya02.botcommands.api.new_components.builder
 
 import com.freya02.botcommands.api.ReceiverConsumer
-import com.freya02.botcommands.api.components.InteractionConstraints
+import com.freya02.botcommands.api.new_components.InteractionConstraints
 import net.dv8tion.jda.api.Permission
 
 interface IConstrainableComponent {
