+++ src/main/kotlin/com/freya02/botcommands/internal/new_components/new/PersistentComponentData.kt
@@ -1,6 +1,6 @@
 package com.freya02.botcommands.internal.new_components.new
 
-import com.freya02.botcommands.api.components.InteractionConstraints
+import com.freya02.botcommands.api.new_components.InteractionConstraints
 import com.freya02.botcommands.internal.new_components.ComponentType
 import com.freya02.botcommands.internal.new_components.LifetimeType
 import com.freya02.botcommands.internal.new_components.PersistentHandler
