+++ src/test/java/com/freya02/botcommands/test/commands/slash/SlashPaginator.java
@@ -4,8 +4,8 @@ import com.freya02.botcommands.api.commands.application.ApplicationCommand;
 import com.freya02.botcommands.api.commands.application.annotations.AppOption;
 import com.freya02.botcommands.api.commands.application.slash.GuildSlashEvent;
 import com.freya02.botcommands.api.commands.application.slash.annotations.JDASlashCommand;
-import com.freya02.botcommands.api.components.InteractionConstraints;
 import com.freya02.botcommands.api.new_components.Components;
+import com.freya02.botcommands.api.new_components.InteractionConstraints;
 import com.freya02.botcommands.api.pagination.paginator.Paginator;
 import com.freya02.botcommands.api.pagination.paginator.PaginatorBuilder;
 import com.freya02.botcommands.api.utils.ButtonContent;
